requirements.txt
pymupdf>=1.25.0
python-docx>=1.1.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.4.0
pdfplumber>=0.11.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
anthropic>=0.28.0
gradio>=4.37.0
python-dotenv>=1.0.0
numpy>=2.0.0